Move slice.go to member.go and rename to valuelocator.go
